KeyLogger-main/requirements.txt
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
idna==3.3
numpy==1.23.1
opencv-python==4.6.0.66
Pillow==9.2.0
pycparser==2.21
pynput==1.7.6
pywin32==304
requests==2.28.1
scipy==1.9.0
six==1.16.0
sounddevice==0.4.4
urllib3==1.26.11
